Move consulting model file to Consulting&SponsorsModels
